Refactor command-related exceptions
