+++ src/main/java/jarvis/command/EventCommand.java
@@ -2,7 +2,7 @@ package jarvis.command;
 
 import java.util.List;
 
-import jarvis.exception.CommandParseException;
+import jarvis.exception.command.CommandParseException;
 import jarvis.storage.Storage;
 import jarvis.task.EventTask;
 import jarvis.task.TaskList;
