+++ src/main/java/jarvis/task/TaskList.java
@@ -5,8 +5,8 @@ import java.util.LinkedList;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import jarvis.exception.InvalidParameterException;
-import jarvis.exception.MissingParameterException;
+import jarvis.exception.command.InvalidParameterException;
+import jarvis.exception.command.MissingParameterException;
 
 /**
  * Container class for tasks and their corresponding operations.
